Trying to work my way through the paper
Figuring out how exactly to structure this all / what operations to do.

Kinda confused about where and how exactly u and l are supposed to be getting updated as we're doing the various different steps here - curosry glance suggested just at the very end, but on closer look that's definitely not true for at least one of them (and possibly BOTH of them).