Step 2 / 7 from the paper!
Take all the points where u(x) is less than s(c(x)).

NOTE: it's NOT u(x)<= s(x). s takes in a CENTROID, not the point itself, so if you pass in the usize index of the point that's wrong. Needs to be the usize of the center that said point is currently assigned to!

Full explanation:
"the upper bound on the distance between it and center it's currently assigned" <=  s(current centroid assigned to x)

(or, <= the distance to the midpoint between centroid c that point x is currently assigned to, and centroid c' which was the centroid exclduing c itself that was closest to c when calculating distances earlier)