Add TODO to prevent unwanted usage
Pretty sure what I'm doing in the training loop isn't _actually_ updating the helpers array (compiler is actually warning that it's fully unused!)

So slapped on a TODO for now. Need to fix that, then it's on to testing + cleanup! (e.g. probably will things out into separate functions for each of the 7 steps. And then maybe test each one... maybe. TBD)