<JUST DID REBASE>
Leaving this as a reminder to find it. Notably there were a couple merge conflicts during the process, and I'm a little worried I may have chosen the wrong piece of code during. (So will want to compare the squashed set of changes *very* carefully to make sure everything outside of Layer is unchanged from master)