Highlight that t=0 for river + preflop
I was confused at first as to why the clustering was "instant" + not actually hitting my TODO. Then I realized, t=0 for the first street that's calculated (river)! See street.rs definition for fn t:

```
    pub const fn t(&self) -> usize {
        match self {
            Self::Pref => 0,
            Self::Flop => crate::KMEANS_FLOP_TRAINING_ITERATIONS,
            Self::Turn => crate::KMEANS_TURN_TRAINING_ITERATIONS,
            Self::Rive => 0,
        }
    }
```

(This is also why there are only definitions for KMEANS_FOO_TRAINING_ITERATIONS for FLOP and TURN over in lib.rs. I.e. they're implicitly 0.)

Given the lack of tests or benchamrks for it, to actually exercise the layer.rs clustering code therefore means waiting for the river to finish up and then for it to move on to the turn (I think).

(Also: figured out how to run this on shortdeck, it's pretty simple: just do `cargo run --features shortdeck`)