Add loss tracking for initialization
This brings this part in line with the pre-existing code by calculating loss / RMS error in the exact same way when it takes the nearest neighbor for all the points

(disclaimer: I don't fully understand the error code! But since this is directly analogous I think it's safe to just replicate it over here. TBD on calculating the loss for the inner steps of the triangle-inequality accelerate code - I'll need to understand the details more before adding stuff there)