rename function again
"step" is confusing because there's steps in the accelerated clustering algorithm. And really what it's doing is in fact getting the 'next' thing each time.

next() as is was too vague imo though, so for now going with 'compute_next_centroids'

(could also just call it next_centroids() I guess... tbd)