Found one possibility for the slowdown
TBD - need to benchmark - but this seems like a very reasonable candidate to investigate further. Since we already computed all the centroid to centroid distances in step 1 - and during step three there's only ever updates to the helpers + reassignment of which centroid is used for specific points. The actual _centroids themselves_ don't get adjusted until step 4, meaning that this computation is completely pointless (it should instead just be a lookup)