WIP - quickly optimizing step 5 (and 6)
Step 5 actually also has a significant runtime for little reason. We should only need to be doing k distance comparisons (each centroid to the new centroid replacing it), but instead we're doing a LOT more than that.

(Not certain this actually will fix it since I'm waiting for the next run, but hese are the only things in step 5 I can see actually being expensive. So I think it's likely to make it go from ~minutes to ~seconds)