Testing out parallelization of helper initialization + tweak logigng
Realized 1. I was mixing up where time was being spent (in self.init() - which is used for both triangle-accelerated and the original version - vs initializing the helpers here) and 2. that I had forgotten to use .par_iter to parallelize this step.

Doing a quick run or two (-F shortdeck) to see if the parallelization actually helps. When parallelized it runs the normal init in ~12 minutes, and then the helper inits in another ~12 minutes. TBD what happens when unparallelized (didn't have fine grained enough logging before) but I think it took quite a bit longer.