Messing with progress bar (TBD: maybe revert)
Seeing if we can get a way to track progress towards initializing these helpers.

(Even without doing this it's a lot faster anyways than before we parallelized, roughly comparable to the first initialization. But it FEELS slow because we're not seeing the progress bar actually move at all.

TBD whether this actually works though / whether we're ok having rayon be a dependency; if not to either, then we may wan to revert this.)