Add inner progress bar for parallelized step 3
Flop clustering is taking 1.5 hours still on step 3 shortdeck, which is a lot / nearly ALL the time spent for it. So will want a way to track that too.
(going to need to potentially play aroudn with whre the progress_count actually should get called - TBD)

Fwiw though - even 1.5h for flop clustering is still a massive improvement! Since non-accelerated this takes ~6 hours per iteration on my machine atm. Just, would like to have visibility into it / to see if we can further speed it up given it's by-far the most expensive part of this new algorithm.