Found another mistake! Forgot to update upper bound
EVERY time we compute d(x,c) OR we update c(x) we're supposed to also update the upper bound. Unlike in Step 3.a, here in Step 3.b wasn't updating upper bound at all (despite having TWO reasons to do so here!)

Could be one potential cause for why this is still taking a while when we scale up to cluster the flop.
That said, even if not, there are a couple other big optimizations to do still.

(E.g. we're cloning the WHOLE hashmap on each loop in step 3 when in theory we should be able to just update just the changed helpers in a cheaper way - possibly entirely in place. TBD.)