(fix prior cl) Actually I slightly misunderstood
We DO want to update it when we've reassigned c(x). BUT, we don't want to reassign on *ever* distance calculation d(x, c) - only whenever calculating d(x, c(x)) specifically.

In practice shouldn't make much of a difference though (and some quick runs suggest that the overall time for step3 doesn't really change as a result; still is about 1.5 minutes)