WIP - testing out lemma 1 for initialization
TBD if this actually improves things yet. (Before this change it took about the same time to initialize these helpers as it did to do all of the other initialization; currently leaving it running to see whether it will speed things up. My guess is that it _probably_ will improve things though)