MASSIVE initialization speedup :D
Now that it's not spamming incorrect error logs we can confirm helper initialization (for my machine, shortdeck) is now around 5x-6x as fast as before! And specifically a fraction of the time it takes to do all of the other initialization in both algorithms (i.e. computing the initial centroids)

This shaves off almost another 1/4 of the total runtime; on a m4 mac mini, turn shortdeck clustering is looking at about:
- 12 minutes to initialize the centroids (same as unaccelerated alg)
- about 2 minutes to initialize the triangle inequality helpers
- about 25 minutes to do all the training iterations (somewhere around a minute each on average)

Also on a side note - there may be additional optimizations we can do, but I think this is now more than good enough of a stopping place. So probably time to go test it, clean it up, and then merge it back into the main branch / replace the unaccelerated algorithm.