Tighter proof of correctness: comparing "loss" at each step
In the mean time until I get a test working this will at least give us SOME insight as to if the results at individual steps probably match. Which might let us avoid waiting another ~week for the unacceelrated fulldeck clustering to finish up.

We probably also should compare the centroids from both to each other at each step in a test, but I've realized that's going to require some heavy refactoring (ie extracting out of layer / passing all the dependencies in via an argument) to do quickly enough for a test.

TBD how much of a performance hit this will be. Likely not much in practice (since O(# points), given it'll only add one emd per point ever) - but it's still not "free" unlike with the unaccelerated version, plus I could be incorrect. So high chacne we'll want to skip these unless explicitly toggled true (or, maybe in non-release mode or something)