Avoid spamming about extra RMS calcs (just do one message since it's basically the entirety of step 4)
Also as a side note - this all worked as hoped so far. The RMS error at each step of turn shortdeck clustering is identical (hooray!) when ran in the same directory after toggling true/false to enable or disable the triangle acceleration. BUT... something I forgot is the error is the same at each step for both algorithms in turn shortdeck clustering. Which to me means: the centroids aren't moving at all after the first iteration in shortdeck! (ie we already converged on the best possible solution we're going to get to after step 1)

So: going to want to also check some of the other choices between (flop shortdeck, turn fulldeck, flop fulldeck) - can't remember which one, but I know for a FACT that at least one of these actually does in fact need additional training steps to convert / that the rms error isn't unchanging at each step.