Add precise tracking for time spent computing RMS (since it's actually pretty big for shortdeck flop!)
Turns out for fulldeck this is a LOT more expensive than shortdeck! (like actually MULTIPLE TIMES MORE than step 3 took, I believe). Decided to report it in aggregate to avoid spamming though - especially since it doesn't really matter how long it takes for an individual centroid vs just how long it takes overall for everything.

That all said - we COULD probably optimize this area a lot by making the top part of step 4 use parallelization (since right now it's all sequential... which is actually my guess for _why_ it took longer than step 3 in this one specific case.)   If so we'll need to restructure the code a bit though to use a .map rather than for loop (both for this new 'rms_calculations_seconds' but also for the per-centroid 'create new centroid that is mean of all points asisgned to it' operation)