RMS needs to be to prior center to match unacclerated closely! (NOT new centers)
After this change, differences in RMS errors are miniscule (something like ~0.001% ish)

___

Noticed that the errors were surprisingly off by a bit, so decided to double check that I wasn't doing something dumb here. Aaaand sure enough - I made a mistake! Unaccelerated is taking the square of the distance to the nearest neighbor centroid *before creating the new set of centroids*. As opposed to afterwards.

As soon as I made the change, error stopped massively varying to becoming super close for turn shortdeck clustering:

- unaccelerated -> 0.03912214 each time
- after-this-commit accelerated -> 0.03912449 each time ( 0.00000235 difference from unaccelerated - so basically matching! especially considering this is all floating point math)
- before-this-commit TRI-accelerated -> 0.037886366 each time (which is quite different from both)

And for flop shortdeck clustering:
- unaccelerated -> 0.05807109 each iteration
- after-this-commit accelerated -> 0.05806913, so => 0.00000196 difference from unacelerated (again hopefully just floating point math issue)
- before-this-commit accelerated -> 0.058941536 each time, which again is noticeably different (not by a huge amount... but still, not ideal)